+++ example/OrchestratorBuildFactoryExample.php
@@ -6,9 +6,11 @@ require __DIR__.'/../vendor/autoload.php';
 require __DIR__.'/autoload.php';
 
 use LDL\File\Directory;
+use LDL\File\Exception\FileException;
+use LDL\File\Helper\DirectoryHelper;
 use LDL\File\Helper\FilePathHelper;
+use LDL\Orchestrator\Builder\Factory\OrchestratorBuilderFactory;
 use LDL\Orchestrator\Console\Command\OrchestratorBuildCommand;
-use LDL\Orchestrator\Loader\OrchestratorLoader;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputDefinition;
@@ -18,8 +20,8 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 $path = FilePathHelper::createAbsolutePath(sys_get_temp_dir(), 'ldl-orchestrator-example');
 try {
     $outputDirectory = Directory::create($path, 0755);
-} catch (\LDL\File\Exception\FileException $e) {
-    \LDL\File\Helper\DirectoryHelper::delete($path);
+} catch (FileException $e) {
+    DirectoryHelper::delete($path);
     $outputDirectory = Directory::create($path, 0755);
 }
 
@@ -58,19 +60,5 @@ $build = new OrchestratorBuildCommand();
 
 $build->execute($argv, new ConsoleOutput());
 
-$orch = OrchestratorLoader::fromJsonFile($outputDirectory->mkpath('ldl-orchestrator-config.json'));
-
-echo "Check some environment variables ...\n\n";
-
-var_dump('ADMIN_APPLICATION_URL', getenv('ADMIN_APPLICATION_URL'));
-var_dump('USER_APPLICATION_URL', getenv('USER_APPLICATION_URL'));
-
-echo "\nPrint out service ID's plus referenced classes ...\n\n";
-
-foreach ($orch->getContainer()->getServiceIDs() as $serviceID) {
-    echo sprintf('%s: %s%s', $serviceID, get_class($orch->getContainer()->get($serviceID)), "\n");
-}
-
-echo "\n";
-
-$outputDirectory->delete();
+$orchBuilder = OrchestratorBuilderFactory::fromJsonFile($outputDirectory->mkpath('ldl-orchestrator-build-config.json'));
+dump($orchBuilder->getConfig()->toArray());