+++ src/Orchestrator/Builder/Config/Exception/OrchestratorBuilderConfigException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LDL\Orchestrator\Builder\Config\Exception;
+
+use LDL\Orchestrator\Builder\Exception\OrchestratorException;
+
+class OrchestratorBuilderConfigException extends OrchestratorException
+{
+}