+++ src/Orchestrator/Builder/Config/Exception/OrchestratorBuilderConfigWriteException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LDL\Orchestrator\Builder\Config\Exception;
+
+class OrchestratorBuilderConfigWriteException extends OrchestratorBuilderConfigException
+{
+}