+++ src/Orchestrator/Builder/Factory/Exception/OrchestratorBuilderFactoryException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LDL\Orchestrator\Builder\Factory\Exception;
+
+use LDL\Orchestrator\Builder\Exception\OrchestratorException;
+
+class OrchestratorBuilderFactoryException extends OrchestratorException
+{
+}