+++ src/Orchestrator/Builder/Factory/OrchestratorBuilderFactory.php
@@ -1,4 +1,7 @@
 <?php
+/**
+ * Contains different methods to create an OrchestratorBuilder.
+ */
 
 declare(strict_types=1);
 
@@ -6,25 +9,28 @@ namespace LDL\Orchestrator\Builder\Factory;
 
 use LDL\DependencyInjection\CompilerPass\Compiler\CompilerPassCompiler;
 use LDL\DependencyInjection\CompilerPass\Finder\CompilerPassFileFinder;
-use LDL\DependencyInjection\CompilerPass\Finder\Options\CompilerPassFileFinderOptions;
+use LDL\DependencyInjection\CompilerPass\Finder\Options\CompilerPassFileFinderOptionsInterface;
 use LDL\DependencyInjection\Container\Builder\LDLContainerBuilder;
 use LDL\DependencyInjection\Service\Compiler\ServiceCompiler;
-use LDL\DependencyInjection\Service\File\Finder\Options\ServiceFileFinderOptions;
+use LDL\DependencyInjection\Service\File\Finder\Options\ServiceFileFinderOptionsInterface;
 use LDL\DependencyInjection\Service\File\Finder\ServiceFileFinder;
 use LDL\Env\Builder\EnvBuilder;
 use LDL\Env\File\Finder\EnvFileFinder;
 use LDL\Env\File\Finder\Options\EnvFileFinderOptionsInterface;
 use LDL\Env\Util\Compiler\EnvCompiler;
 use LDL\Env\Util\File\Parser\EnvFileParser;
+use LDL\Orchestrator\Builder\Config\OrchestratorBuilderConfig;
+use LDL\Orchestrator\Builder\Config\OrchestratorBuilderConfigInterface;
+use LDL\Orchestrator\Builder\Factory\Exception\OrchestratorBuilderFactoryException;
 use LDL\Orchestrator\Builder\OrchestratorBuilder;
 use LDL\Orchestrator\Builder\OrchestratorBuilderInterface;
 
-class OrchestratorBuilderFactory
+class OrchestratorBuilderFactory implements OrchestratorBuilderFactoryInterface
 {
     public static function create(
         EnvFileFinderOptionsInterface $envFileFinderOptions,
-        ServiceFileFinderOptions $serviceFileFinderOptions,
-        CompilerPassFileFinderOptions $compilerPassFileFinderOptions
+        ServiceFileFinderOptionsInterface $serviceFileFinderOptions,
+        CompilerPassFileFinderOptionsInterface $compilerPassFileFinderOptions
     ): OrchestratorBuilderInterface {
         $envFileFinder = new EnvFileFinder($envFileFinderOptions);
         $envCompiler = new EnvCompiler();
@@ -45,5 +51,38 @@ class OrchestratorBuilderFactory
             $envBuilder,
             $containerBuilder
         );
+    }
+
+    public static function fromJsonFile(string $file): OrchestratorBuilderInterface
+    {
+        try {
+            return self::fromConfig(OrchestratorBuilderConfig::fromJsonFile($file));
+        } catch (\Throwable $e) {
+            $msg = sprintf('Unable to create OrchestratorBuilder from file %s', $file);
+            throw new OrchestratorBuilderFactoryException($msg, 0, $e);
+        }
+    }
+
+    public static function fromJsonString(string $json)
+    {
+        try {
+            return self::fromConfig(OrchestratorBuilderConfig::fromJsonString($json));
+        } catch (\Throwable $e) {
+            $msg = 'Unable to create OrchestratorBuilder from JSON string';
+            throw new OrchestratorBuilderFactoryException($msg, 0, $e);
+        }
+    }
+
+    public static function fromConfig(OrchestratorBuilderConfigInterface $config): OrchestratorBuilderInterface
+    {
+        try {
+            return self::create(
+                $config->getEnvFileFinderOptions(),
+                $config->getServiceFileFinderOptions(),
+                $config->getCompilerPassFileFinderOptions()
+            );
+        } catch (\Throwable $e) {
+            throw new OrchestratorBuilderFactoryException('Could not create orchestrator builder from config', 0, $e);
+        }
     }
 }