+++ src/Orchestrator/Config/OrchestratorConfigInterface.php
@@ -7,14 +7,12 @@ namespace LDL\Orchestrator\Config;
 use LDL\File\Collection\ReadableFileCollection;
 use LDL\File\Contracts\FileInterface;
 use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Framework\Base\Contracts\JsonFactoryInterface;
+use LDL\Framework\Base\Contracts\JsonFileFactoryInterface;
 use LDL\Framework\Base\Contracts\Type\ToArrayInterface;
 
-interface OrchestratorConfigInterface extends \JsonSerializable, ToArrayInterface, ArrayFactoryInterface
+interface OrchestratorConfigInterface extends \JsonSerializable, ToArrayInterface, ArrayFactoryInterface, JsonFileFactoryInterface, JsonFactoryInterface
 {
-    public static function fromJSONString(string $json): OrchestratorConfigInterface;
-
-    public static function fromJSONFile(string $file): OrchestratorConfigInterface;
-
     public function getContainerFile(): FileInterface;
 
     public function getServiceFiles(): ReadableFileCollection;
@@ -24,4 +22,6 @@ interface OrchestratorConfigInterface extends \JsonSerializable, ToArrayInterfac
     public function getEnvFiles(): ReadableFileCollection;
 
     public function getCompilerPassFiles(): ReadableFileCollection;
+
+    public function write(string $path, bool $force): FileInterface;
 }