+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# LDL Env Util Changelog
+# LDL Orchestrator Changelog
 
 All changes to this project are documented in this file.
 
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- feature/1201943740776108 - Separate OrchestratorBuilder config from main OrchestratorConfig
 - feature/1201611356837883 - Integrate ldl-env-builder + ldl-container-builder
 
 ### Changed